+++ homework_board/apps/db/lib/db/assignments.ex
@@ -1,8 +1,8 @@
-defmodule Db.Assignment do
+defmodule Db.Assignments do
   use Ecto.Schema
   require Ecto.Query
 
-  schema "assignment" do
+  schema "assignments" do
     field :title, :string
     field :description, :string
     field :priority, :integer
@@ -11,31 +11,33 @@ defmodule Db.Assignment do
   def changeset(assignment, params \\ %{}) do
     assignment
     |> Ecto.Changeset.cast(params, [:title, :description, :priority])
-    |> Ecto.Changeset.validate_required([:title, :description, :priority])
+    |> Ecto.Changeset.validate_required([:title])
   end
 
   def get_assignment() do
-    Db.Assignment
+    Db.Assignments
     |> Db.Repo.all
   end
 
   def get_assignment(id) when is_integer(id) do
-    Db.Assignment
+    Db.Assignments
     |> Db.Repo.get(id)
   end
 
   def add_assignment(assignment) do
-    changeset = Db.Assignment.changeset(assignment, %{})
+    changeset = Db.Assignments.changeset(assignment, %{})
     Db.Repo.insert(changeset)
     |> handle_query_response
   end
 
   def get_all_with_priority(priority) when is_integer(priority) do
-    Db.Assignment |> Ecto.Query.where(priority: ^priority) |> Db.Repo.all
+    Db.Assignments
+    |> Ecto.Query.where(priority: ^priority)
+    |> Db.Repo.all
   end
 
   def update_assignment(assignment, newAssignment) do
-    changeset = Db.Assignment.changeset(assignment, newAssignment)
+    changeset = Db.Assignments.changeset(assignment, newAssignment)
     Db.Repo.update(changeset)
     |> handle_query_response
   end
@@ -48,9 +50,15 @@ defmodule Db.Assignment do
   defp handle_query_response({ :ok, _changeset } = _response) do
     IO.puts "Successfully modified database"
   end
-  defp handle_query_response({ :error, _changeset } = _response) do
-    # Is raising an error proper in this situatuon?
-    raise "Error modifying database"
+  defp handle_query_response({ :error, changeset } = _response) do
+    # Is raising an error proper in this situation?
+    errors = Ecto.Changeset.traverse_errors(changeset, fn {msg, opts} ->
+      Enum.reduce(opts, msg, fn {key, value}, acc ->
+        String.replace(acc, "%{#{key}}", to_string(value))
+      end)
+    end)
+
+    raise ("Error modifying database: #{inspect errors}")
   end
 
 end
