+++ homework_board/apps/frontend/lib/frontend_web/controllers/new_assignment_controller.ex
@@ -2,6 +2,19 @@ defmodule FrontendWeb.NewAssignmentController do
   use FrontendWeb, :controller
 
   def index(conn, _params) do
-    render conn, "index.html"
+    changeset = Db.Assignments.changeset(%Db.Assignments{})
+    conn
+    |> assign(:changeset, changeset)
+    |> render("index.html")
   end
+
+  def create(conn, %{"assignments" => %{"title" => title, "priority" => priority_str}} = _params) do
+    {priority, _} = Integer.parse(priority_str)
+
+    %Db.Assignments{title: title, priority: priority}
+    |> Db.Assignments.add_assignment()
+
+    index(conn, %{})
+  end
+
 end
